:recycle: [#1068] -- bring dynamic config mutations into main component/registry
This deletes the dynamic config-specific registry, while mostly keeping
the public API intact.
